Use crop function istead of bySegments, added Iterator and Option variant to read by segments, code refactor.
