+++ raster/src/main/scala/geotrellis/raster/GridBounds.scala
@@ -211,6 +211,7 @@ case class GridBounds(colMin: Int, rowMin: Int, colMax: Int, rowMax: Int) {
       )
     }
 
+  /** Return the union of GridBounds. */
   def combine(other: GridBounds): GridBounds =
     GridBounds(
       colMin = math.min(this.colMin, other.colMin),
@@ -219,7 +220,12 @@ case class GridBounds(colMin: Int, rowMin: Int, colMax: Int, rowMax: Int) {
       rowMax = math.max(this.rowMax, other.rowMax)
     )
 
+  /** Empty gridbounds contain nothing, though non empty gridbounds contains iteslf */
   def contains(other: GridBounds): Boolean =
-    colMin <= other.colMin && rowMin <= other.rowMin &&
-    colMax >= other.colMax && rowMax >= other.rowMax
+    if(colMin == 0 && colMax == 0 && rowMin == 0 && rowMax == 0) false
+    else
+      other.colMin >= colMin &&
+      other.rowMin >= rowMin &&
+      other.colMax <= colMax &&
+      other.rowMax <= rowMax
 }
