+++ spark/src/main/scala/geotrellis/spark/io/RasterReader.scala
@@ -16,19 +16,18 @@
 
 package geotrellis.spark.io
 
-import java.time.format.DateTimeFormatter
-import java.time.{ZoneOffset, ZonedDateTime}
-
 import geotrellis.proj4._
 import geotrellis.raster._
 import geotrellis.raster.io.geotiff._
 import geotrellis.raster.io.geotiff.reader.GeoTiffReader
 import geotrellis.spark._
-import geotrellis.spark.tiling.MapKeyTransform
 import geotrellis.util.{ByteReader, StreamingByteReader}
-import geotrellis.vector.{Extent, ProjectedExtent}
+import geotrellis.vector.ProjectedExtent
+
 import spire.syntax.cfor._
 
+import java.time.format.DateTimeFormatter
+import java.time.{ZoneOffset, ZonedDateTime}
 
 /**
   * Type class to read a raster either fully or partially from a ByteReader.
@@ -42,7 +41,11 @@ import spire.syntax.cfor._
 trait RasterReader[-O, R] extends Serializable {
   def readFully(byteReader: ByteReader, options: O): R
   def readWindow(byteReader: StreamingByteReader, pixelWindow: GridBounds, options: O): R
-  def readSegments(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: O): Iterator[R]
+
+  /** Read segments iterator, each segment into a single R */
+  def readSegmentsIterator(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: O): Iterator[R]
+  /** Reads all segments into one R */
+  def readSegments(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: O): Option[R]
 }
 
 object RasterReader {
@@ -82,9 +85,6 @@ object RasterReader {
     result.toArray
   }
 
-  def mapTransform(info: GeoTiffReader.GeoTiffInfo) =
-    MapKeyTransform(info.extent, info.segmentLayout.totalCols, info.segmentLayout.totalRows)
-
   implicit def singlebandGeoTiffReader = new RasterReader[Options, (ProjectedExtent, Tile)] {
     def readFully(byteReader: ByteReader, options: Options) = {
       val geotiff = SinglebandGeoTiff(byteReader)
@@ -98,13 +98,25 @@ object RasterReader {
       (ProjectedExtent(raster.extent, options.crs.getOrElse(geotiff.crs)), raster.tile)
     }
 
-    def readSegments(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
+    def readSegmentsIterator(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
       val geoTiff = GeoTiffReader.geoTiffSinglebandTile(info)
       val gridBounds = geoTiff.gridBounds
       gbs
         .filter(gridBounds.contains)
-        .map { gb => (ProjectedExtent(mapTransform(info)(gb), options.crs.getOrElse(info.crs)), geoTiff.crop(gb)) }
+        .map { gb => (ProjectedExtent(info.mapTransform(gb), options.crs.getOrElse(info.crs)), geoTiff.crop(gb)) }
         .toIterator
+    }
+
+    def readSegments(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
+      val geoTiff = GeoTiffReader.geoTiffSinglebandTile(info)
+      val gridBounds = geoTiff.gridBounds
+      val filteredGbs = gbs.filter(gridBounds.contains)
+      if(filteredGbs.nonEmpty) {
+        val gb = filteredGbs.reduce(_ combine _)
+        val extent = info.mapTransform(gb)
+
+        Some(ProjectedExtent(extent, options.crs.getOrElse(info.crs)) -> geoTiff.crop(gb))
+      } else None
     }
   }
 
@@ -121,13 +133,25 @@ object RasterReader {
       (ProjectedExtent(raster.extent, options.crs.getOrElse(geotiff.crs)), raster.tile)
     }
 
-    def readSegments(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
+    def readSegmentsIterator(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
       val geoTiff = GeoTiffReader.geoTiffMultibandTile(info)
       val gridBounds = geoTiff.gridBounds
       gbs
         .filter(gridBounds.contains)
-        .map { gb => (ProjectedExtent(mapTransform(info)(gb), options.crs.getOrElse(info.crs)), geoTiff.crop(gb)) }
+        .map { gb => (ProjectedExtent(info.mapTransform(gb), options.crs.getOrElse(info.crs)), geoTiff.crop(gb)) }
         .toIterator
+    }
+
+    def readSegments(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
+      val geoTiff = GeoTiffReader.geoTiffMultibandTile(info)
+      val gridBounds = geoTiff.gridBounds
+      val filteredGbs = gbs.filter(gridBounds.contains)
+      if(filteredGbs.nonEmpty) {
+        val gb = filteredGbs.reduce(_ combine _)
+        val extent = info.mapTransform(gb)
+
+        Some(ProjectedExtent(extent, options.crs.getOrElse(info.crs)) -> geoTiff.crop(gb))
+      } else None
     }
   }
 
@@ -148,18 +172,30 @@ object RasterReader {
       (TemporalProjectedExtent(raster.extent, crs, time), raster.tile)
     }
 
-    def readSegments(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
+    def readSegmentsIterator(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
       val geoTiff = GeoTiffReader.geoTiffSinglebandTile(info)
       val gridBounds = geoTiff.gridBounds
       gbs
         .filter(gridBounds.contains)
         .map { gb =>
           (TemporalProjectedExtent(
-            extent = info.extent,
+            info.mapTransform(gb),
             crs = options.crs.getOrElse(info.crs),
             options.parseTime(info.tags)), geoTiff.crop(gb))
         }
         .toIterator
+    }
+
+    def readSegments(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
+      val geoTiff = GeoTiffReader.geoTiffSinglebandTile(info)
+      val gridBounds = geoTiff.gridBounds
+      val filteredGbs = gbs.filter(gridBounds.contains)
+      if(filteredGbs.nonEmpty) {
+        val gb = filteredGbs.reduce(_ combine _)
+        val extent = info.mapTransform(gb)
+
+        Some(TemporalProjectedExtent(extent, options.crs.getOrElse(info.crs), options.parseTime(info.tags)) -> geoTiff.crop(gb))
+      } else None
     }
   }
 
@@ -180,18 +216,30 @@ object RasterReader {
       (TemporalProjectedExtent(raster.extent, crs, time), raster.tile)
     }
 
-    def readSegments(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
+    def readSegmentsIterator(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
       val geoTiff = GeoTiffReader.geoTiffMultibandTile(info)
       val gridBounds = geoTiff.gridBounds
       gbs
         .filter(gridBounds.contains)
         .map { gb =>
           (TemporalProjectedExtent(
-            mapTransform(info)(gb),
+            info.mapTransform(gb),
             options.crs.getOrElse(info.crs),
             options.parseTime(info.tags)), geoTiff.crop(gb))
         }
         .toIterator
+    }
+
+    def readSegments(gbs: Array[GridBounds], info: GeoTiffReader.GeoTiffInfo, options: Options) = {
+      val geoTiff = GeoTiffReader.geoTiffMultibandTile(info)
+      val gridBounds = geoTiff.gridBounds
+      val filteredGbs = gbs.filter(gridBounds.contains)
+      if(filteredGbs.nonEmpty) {
+        val gb = filteredGbs.reduce(_ combine _)
+        val extent = info.mapTransform(gb)
+
+        Some(TemporalProjectedExtent(extent, options.crs.getOrElse(info.crs), options.parseTime(info.tags)) -> geoTiff.crop(gb))
+      } else None
     }
   }
 }
